Refactor AccessTokenAuthenticationFilterTest and move CustomUserDetails
AccessTokenAuthenticationFilterTest has been refactored to improve readability and efficiency. Tests are focused on different return cases such as token expiration and access denial. The `CustomUserDetails` class has been moved from `greencity.controller` package to `greencity.security.model` package for better organization of code.
