+++ core/src/main/java/greencity/security/model/CustomUserDetails.java
@@ -1,25 +1,27 @@
-package greencity.controller;
+package greencity.security.model;
 
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 
 import java.util.Collection;
 
+@Getter
+@Setter
+@EqualsAndHashCode(callSuper = true)
 public class CustomUserDetails extends User {
     private final Long id;
     private String email;
 
-    public CustomUserDetails(String email, String password, Collection<? extends GrantedAuthority> authorities, Long id) {
+    public CustomUserDetails(
+            String email,
+            String password,
+            Collection<? extends GrantedAuthority> authorities,
+            Long id) {
         super(email, password, authorities);
         this.id = id;
         this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
     }
 }
