+++ core/src/main/java/greencity/security/service/UserDetailsServiceImpl.java
@@ -1,6 +1,6 @@
 package greencity.security.service;
 
-import greencity.controller.CustomUserDetails;
+import greencity.security.model.CustomUserDetails;
 import greencity.entity.User;
 import greencity.repository.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
